Fix multiple popup subscriptions for embedded context
